move useRemoveSongFromLibrary()
