+++ projects/graphql-client/src/index.ts
@@ -28,3 +28,4 @@ export * from "./remote/mutations/useInviteToShare"
 export * from "./remote/mutations/useRenameShare"
 export * from "./remote/mutations/useRenamePlaylist"
 export * from "./remote/mutations/useRemoveSongsFromPlaylist"
+export * from "./remote/mutations/useRemoveSongFromLibrary"
